requirements.txt
argparse==1.2.1
boto==2.30.0
django-boto==0.3.1
django-storages==1.1.8
mock==1.0.1
wsgiref==0.1.2
